move bindings to module
